directory cleaner now is able to clean the directory recursively. be carefull using it!

+++ db-bench/src/main/java/com/eris4/benchdb/database/common/DirectoryCleaner.java
@@ -11,9 +11,13 @@ public class DirectoryCleaner {
 	public static void clean(String directoryName){
 		try {
 			for (File file : new File(directoryName).listFiles()) {
-				boolean b = file.delete();
-				if (b==false)
-					throw new FileBusyException("cannot delete the file");
+				if (file.isDirectory()){
+					clean(file.getAbsolutePath()) ;
+				}else {
+					boolean b = file.delete();
+					if (b==false)
+						throw new FileBusyException("cannot delete the file");
+				}				
 			}
 		} catch (NullPointerException e) {
 			logger.warn("Null Pointer Exception while deleting the db "+directoryName+". Probably the first time you run this test! Don't worry");
